[refactor] 캐시 서비스 리팩토링
- JsonConverter를 사용해 JSON <-> 객체 변환
- RedisKeyExtractor를 사용해 Redis에 사용되는 키를 id로부터 추출
